fix: preserve escaped/entity <br> text in cells and split table extension
Review follow-ups:

- Markdown fidelity: the cell <br>-to-hard-break conversion now runs before
  markdown-it's text_join rule, where escaped (\<br>) and entity (&lt;br&gt;)
  forms are still separate text_special tokens — only raw <br> in plain text
  converts to a line break. On the way out, literal < in cell text is escaped
  via the serializer's escapeExtraCharacters so it cannot be reinterpreted as
  a break; code spans keep their raw form on both sides. Covered by new
  round-trip tests (escaped, entity, and code-span cases).
- Split the 1,000-line table extension into focused modules under
  src/table/: table-config (shared types/defaults), table-schema (node
  specs), table-commands (commands, navigation, queries), table-markdown
  (tokenizer, parse specs, serializer — the save-path module), and
  table-plugins (keymap, active-cell decoration, fixTables). setupTable()
  stays as composition in table/index.ts; the public API is unchanged.
- Wired Escape into the table-menu plugin so the floating menu can be
  dismissed until the document changes; the previously dead dismissal state
  and command are now live, with an e2e assertion.
- Removed the stray blank line at the end of the plan file.
