+++ packages/js-lib/banger-editor/src/table/table-commands.ts
@@ -1,10 +1,3 @@
-import type { MarkdownSerializerState } from 'prosemirror-markdown';
-import {
-  type CollectionType,
-  collection,
-  keybinding,
-  PRIORITY,
-} from './common';
 import {
   addColumnAfter,
   addColumnBefore,
@@ -12,217 +5,33 @@ import {
   CellSelection,
   type Command,
   cellAround,
-  chainCommands,
-  Decoration,
-  DecorationSet,
   deleteColumn,
   deleteRow,
   deleteTable,
   type EditorState,
-  fixTables,
   goToNextCell,
   isInTable,
-  type NodeType,
   nextCell,
-  Plugin,
-  PluginKey,
   type PMNode,
-  type Schema,
   selectedRect,
   selectionCell,
   TableMap,
   TextSelection,
-  tableEditing,
-  tableNodes,
-} from './pm';
+} from '../pm';
 import {
-  defaultGetParagraphNodeType,
   findParentNodeOfType,
   getNodeType,
   insertEmptyParagraphAboveNode,
   insertEmptyParagraphBelowNode,
   safeInsert,
-} from './pm-utils';
+} from '../pm-utils';
+import {
+  parseAlign,
+  type RequiredConfig,
+  type TableCellAlign,
+} from './table-config';
 
-export type TableCellAlign = 'left' | 'center' | 'right';
-
-const ALIGN_VALUES: readonly TableCellAlign[] = ['left', 'center', 'right'];
-
-export type TableConfig = {
-  tableGroup?: string;
-  defaultRows?: number;
-  defaultColumns?: number;
-  getParagraphNodeType?: (schema: Schema) => NodeType;
-  hardBreakNodeName?: string;
-  // keys
-  keyGoToNextCell?: string | false;
-  keyGoToPrevCell?: string | false;
-};
-
-type RequiredConfig = Required<TableConfig>;
-
-const DEFAULT_CONFIG: RequiredConfig = {
-  tableGroup: 'block',
-  defaultRows: 3,
-  defaultColumns: 3,
-  getParagraphNodeType: defaultGetParagraphNodeType,
-  hardBreakNodeName: 'hard_break',
-  keyGoToNextCell: 'Tab',
-  keyGoToPrevCell: 'Shift-Tab',
-};
-
-export function setupTable(userConfig?: TableConfig) {
-  const config = {
-    ...DEFAULT_CONFIG,
-    ...userConfig,
-  };
-
-  const nodes = tableNodes({
-    tableGroup: config.tableGroup,
-    // Inline-only cells keep the document representable as Markdown pipe
-    // tables; block content inside cells has no faithful pipe-table form.
-    cellContent: 'inline*',
-    cellAttributes: {
-      align: {
-        default: null,
-        getFromDOM: (dom) => parseAlign(dom.style.textAlign),
-        setDOMAttr: (value, attrs) => {
-          const align = parseAlign(value);
-          if (align) {
-            attrs.style = `${attrs.style ?? ''}text-align: ${align};`;
-          }
-        },
-      },
-    },
-  });
-
-  // Inline-content cells are textblocks, which makes the positions between
-  // cells valid gap-cursor spots by prosemirror-gapcursor's default rule.
-  // A gap cursor inside a row is never meaningful, so forbid it.
-  nodes.table_row = { ...nodes.table_row, allowGapCursor: false };
-
-  const plugin = {
-    keybindings: pluginKeybindings(config),
-    tableEditing: tableEditing(),
-    fixTables: pluginFixTables(),
-    activeCell: pluginActiveCell(),
-  };
-
-  const command = {
-    insertTable: insertTable(config),
-    addRowAbove: addRowAbove(),
-    addRowBelow: addRowBelow(),
-    deleteRow: deleteTableRow(),
-    addColumnLeft: addColumnLeft(),
-    addColumnRight: addColumnRight(),
-    deleteColumn: deleteTableColumn(),
-    deleteTable: deleteWholeTable(),
-    goToNextCell: goToNextCellOrExtend(),
-    goToPrevCell: goToPrevCell(),
-    goToRowBelow: goToRowBelow(),
-    setColumnAlign,
-  };
-
-  return collection({
-    id: 'table',
-    nodes,
-    plugin,
-    command,
-    query: {
-      isTableActive,
-      activeTableCell,
-    },
-    markdown: markdown(),
-  });
-}
-
-// PLUGINS
-function pluginKeybindings(config: RequiredConfig) {
-  return () => {
-    return keybinding(
-      [
-        [config.keyGoToNextCell, goToNextCellOrExtend()],
-        [config.keyGoToPrevCell, goToPrevCell()],
-        // Line breaks inside cells persist as <br>, the GFM convention for
-        // multi-line pipe-table cells.
-        ['Enter', insertLineBreakInCell(config)],
-        ['Shift-Enter', insertLineBreakInCell(config)],
-        ['Mod-Enter', exitTableBelow(config)],
-        [
-          'ArrowUp',
-          chainCommands(
-            moveRowVertical('up', config),
-            enterTableVertical('up'),
-          ),
-        ],
-        [
-          'ArrowDown',
-          chainCommands(
-            moveRowVertical('down', config),
-            enterTableVertical('down'),
-          ),
-        ],
-        ['ArrowLeft', moveCellHorizontal(-1, config)],
-        ['ArrowRight', moveCellHorizontal(1, config)],
-        [
-          'Delete',
-          chainCommands(
-            deleteTableOnFullCellSelection(),
-            deleteEmptyBlockNextToTable('before'),
-          ),
-        ],
-        [
-          'Backspace',
-          chainCommands(
-            deleteTableOnFullCellSelection(),
-            deleteEmptyBlockNextToTable('after'),
-          ),
-        ],
-      ],
-      'table',
-      PRIORITY.high,
-    );
-  };
-}
-
-function pluginActiveCell() {
-  const key = new PluginKey('table-active-cell');
-  // Highlights the cell the cursor is in, like other Notion-style editors,
-  // so the typing target is obvious.
-  return new Plugin({
-    key,
-    props: {
-      decorations(state) {
-        if (!state.selection.empty || !isInTable(state)) {
-          return null;
-        }
-        const $cell = cellAround(state.selection.$from);
-        const cell = $cell?.nodeAfter;
-        if (!$cell || !cell) {
-          return null;
-        }
-        return DecorationSet.create(state.doc, [
-          Decoration.node($cell.pos, $cell.pos + cell.nodeSize, {
-            class: 'prosemirror-active-table-cell',
-          }),
-        ]);
-      },
-    },
-  });
-}
-
-function pluginFixTables() {
-  // Repairs structurally invalid tables (e.g. after a partial paste) so the
-  // serializer never sees a non-rectangular table.
-  return new Plugin({
-    key: new PluginKey('table-fix-tables'),
-    appendTransaction: (_transactions, oldState, newState) =>
-      fixTables(newState, oldState),
-  });
-}
-
-// COMMANDS
-function insertTable(config: RequiredConfig) {
+export function insertTable(config: RequiredConfig) {
   return ({
     rows = config.defaultRows,
     columns = config.defaultColumns,
@@ -278,7 +87,7 @@ function findTablePos(doc: PMNode, tableNode: PMNode): number | null {
   return found;
 }
 
-function addRowAbove(): Command {
+export function addRowAbove(): Command {
   return (state, dispatch) => {
     if (!isInTable(state)) {
       return false;
@@ -296,7 +105,7 @@ function addRowAbove(): Command {
   };
 }
 
-function addRowBelow(): Command {
+export function addRowBelow(): Command {
   return (state, dispatch) => {
     if (!isInTable(state)) {
       return false;
@@ -309,27 +118,27 @@ function addRowBelow(): Command {
   };
 }
 
-function deleteTableRow(): Command {
+export function deleteTableRow(): Command {
   return deleteRow;
 }
 
-function addColumnLeft(): Command {
+export function addColumnLeft(): Command {
   return addColumnBefore;
 }
 
-function addColumnRight(): Command {
+export function addColumnRight(): Command {
   return addColumnAfter;
 }
 
-function deleteTableColumn(): Command {
+export function deleteTableColumn(): Command {
   return deleteColumn;
 }
 
-function deleteWholeTable(): Command {
+export function deleteWholeTable(): Command {
   return deleteTable;
 }
 
-function goToNextCellOrExtend(): Command {
+export function goToNextCellOrExtend(): Command {
   return (state, dispatch) => {
     if (!isInTable(state)) {
       return false;
@@ -358,7 +167,7 @@ function goToNextCellOrExtend(): Command {
   };
 }
 
-function goToPrevCell(): Command {
+export function goToPrevCell(): Command {
   return (state, dispatch) => {
     if (!isInTable(state)) {
       return false;
@@ -371,7 +180,7 @@ function goToPrevCell(): Command {
   };
 }
 
-function goToRowBelow(): Command {
+export function goToRowBelow(): Command {
   return (state, dispatch) => {
     if (!isInTable(state) || !state.selection.empty) {
       return false;
@@ -394,7 +203,7 @@ function goToRowBelow(): Command {
 }
 
 /** Enter/Shift-Enter inside a cell insert a line break (persisted as <br>). */
-function insertLineBreakInCell(config: RequiredConfig): Command {
+export function insertLineBreakInCell(config: RequiredConfig): Command {
   return (state, dispatch) => {
     if (!isInTable(state)) {
       return false;
@@ -413,7 +222,7 @@ function insertLineBreakInCell(config: RequiredConfig): Command {
 }
 
 /** Mod-Enter leaves the table downward, like exiting a code block. */
-function exitTableBelow(config: RequiredConfig): Command {
+export function exitTableBelow(config: RequiredConfig): Command {
   return (state, dispatch) => {
     if (!isInTable(state)) {
       return false;
@@ -426,7 +235,7 @@ function exitTableBelow(config: RequiredConfig): Command {
  * Backspace/Delete on a cell selection that covers the whole table deletes
  * the table itself instead of just emptying every cell.
  */
-function deleteTableOnFullCellSelection(): Command {
+export function deleteTableOnFullCellSelection(): Command {
   return (state, dispatch) => {
     if (!(state.selection instanceof CellSelection)) {
       return false;
@@ -455,7 +264,7 @@ function deleteTableOnFullCellSelection(): Command {
  * browser's native caret motion fights the isolating cell boundaries and
  * the cursor gets stuck, so the whole vertical move is owned here.
  */
-function moveRowVertical(
+export function moveRowVertical(
   direction: 'up' | 'down',
   config: RequiredConfig,
 ): Command {
@@ -492,7 +301,7 @@ function moveRowVertical(
  * caret motion cannot cross the table's isolating boundary, so without this
  * the cursor gets stuck next to the table.
  */
-function enterTableVertical(direction: 'up' | 'down'): Command {
+export function enterTableVertical(direction: 'up' | 'down'): Command {
   return (state, dispatch, view) => {
     if (!state.selection.empty || isInTable(state)) {
       return false;
@@ -554,7 +363,7 @@ function enterTableVertical(direction: 'up' | 'down'): Command {
  * join commands cannot cross the table's isolating boundary and leave the
  * document unchanged.
  */
-function deleteEmptyBlockNextToTable(side: 'before' | 'after'): Command {
+export function deleteEmptyBlockNextToTable(side: 'before' | 'after'): Command {
   return (state, dispatch) => {
     if (!state.selection.empty || isInTable(state)) {
       return false;
@@ -685,7 +494,10 @@ function exitTable(direction: 'up' | 'down', config: RequiredConfig): Command {
  * or next cell in reading order (wrapping across rows), and leaves the table
  * from the very first or last cell.
  */
-function moveCellHorizontal(dir: -1 | 1, config: RequiredConfig): Command {
+export function moveCellHorizontal(
+  dir: -1 | 1,
+  config: RequiredConfig,
+): Command {
   return (state, dispatch) => {
     if (!isInTable(state)) {
       return false;
@@ -748,7 +560,7 @@ function moveCellHorizontal(dir: -1 | 1, config: RequiredConfig): Command {
   };
 }
 
-function setColumnAlign(align: TableCellAlign | null): Command {
+export function setColumnAlign(align: TableCellAlign | null): Command {
   return (state, dispatch) => {
     if (!isInTable(state)) {
       return false;
@@ -780,7 +592,7 @@ function setColumnAlign(align: TableCellAlign | null): Command {
 }
 
 // QUERIES
-function isTableActive(state: EditorState): boolean {
+export function isTableActive(state: EditorState): boolean {
   return isInTable(state);
 }
 
@@ -792,7 +604,9 @@ export type ActiveTableCell = {
   align: TableCellAlign | null;
 };
 
-function activeTableCell(state: EditorState): ActiveTableCell | undefined {
+export function activeTableCell(
+  state: EditorState,
+): ActiveTableCell | undefined {
   if (!isInTable(state)) {
     return undefined;
   }
@@ -805,232 +619,4 @@ function activeTableCell(state: EditorState): ActiveTableCell | undefined {
     isHeaderRow: rect.top === 0,
     align: parseAlign($cell.nodeAfter?.attrs.align),
   };
-}
-
-// MARKDOWN
-function markdown(): CollectionType['markdown'] {
-  return {
-    nodes: {
-      table: {
-        toMarkdown: tableToMarkdown,
-        parseMarkdown: {
-          table: { block: 'table' },
-          thead: { ignore: true },
-          tbody: { ignore: true },
-        },
-      },
-      table_row: {
-        // Rows are serialized by the table node above.
-        toMarkdown: () => {},
-        parseMarkdown: {
-          tr: { block: 'table_row' },
-        },
-      },
-      table_header: {
-        toMarkdown: () => {},
-        parseMarkdown: {
-          th: {
-            block: 'table_header',
-            getAttrs: (tok) => ({
-              align: alignFromToken(tok.attrGet('style')),
-            }),
-          },
-        },
-      },
-      table_cell: {
-        toMarkdown: () => {},
-        parseMarkdown: {
-          td: {
-            block: 'table_cell',
-            getAttrs: (tok) => ({
-              align: alignFromToken(tok.attrGet('style')),
-            }),
-          },
-        },
-      },
-    },
-    tokenizerPlugins: [
-      (md) => {
-        md.enable('table');
-        // GFM cells hold line breaks as literal <br>. The tokenizer runs
-        // with html disabled, so <br> arrives as plain text; convert it to
-        // hardbreak tokens, but only inside table cells so <br> text in
-        // normal paragraphs keeps round-tripping as text.
-        md.core.ruler.push('table-cell-br', (state) => {
-          let cellDepth = 0;
-          for (const token of state.tokens) {
-            if (token.type === 'th_open' || token.type === 'td_open') {
-              cellDepth++;
-            } else if (token.type === 'th_close' || token.type === 'td_close') {
-              cellDepth--;
-            } else if (
-              cellDepth > 0 &&
-              token.type === 'inline' &&
-              token.children
-            ) {
-              token.children = splitBrIntoHardbreaks(
-                token.children,
-                state.Token,
-              );
-            }
-          }
-        });
-      },
-    ],
-  };
-}
-
-const CELL_BR_RE = /<br\s*\/?>/i;
-
-function splitBrIntoHardbreaks<
-  T extends { type: string; content: string; level: number },
->(children: T[], TokenCtor: new (type: string, tag: string, nesting: 0) => T) {
-  const result: T[] = [];
-  for (const child of children) {
-    if (child.type !== 'text' || !CELL_BR_RE.test(child.content)) {
-      result.push(child);
-      continue;
-    }
-    const parts = child.content.split(new RegExp(CELL_BR_RE.source, 'gi'));
-    parts.forEach((part, index) => {
-      if (part) {
-        const text = new TokenCtor('text', '', 0);
-        text.content = part;
-        text.level = child.level;
-        result.push(text);
-      }
-      if (index < parts.length - 1) {
-        const br = new TokenCtor('hardbreak', 'br', 0);
-        br.level = child.level;
-        result.push(br);
-      }
-    });
-  }
-  return result;
-}
-
-function parseAlign(value: unknown): TableCellAlign | null {
-  return ALIGN_VALUES.find((align) => align === value) ?? null;
-}
-
-function alignFromToken(style: string | null): TableCellAlign | null {
-  const match = style?.match(/text-align:\s*(left|center|right)/);
-  return parseAlign(match?.[1]);
-}
-
-function tableToMarkdown(state: MarkdownSerializerState, node: PMNode) {
-  // Flush pending block separation now, so the captured cell output below
-  // cannot swallow the "\n\n" that belongs between the previous block and
-  // this table.
-  state.write();
-
-  const rows: string[][] = [];
-  const aligns: (TableCellAlign | null)[] = [];
-
-  node.forEach((row, _offset, rowIndex) => {
-    const cells: string[] = [];
-    row.forEach((cell) => {
-      if (rowIndex === 0) {
-        aligns.push(parseAlign(cell.attrs.align));
-      }
-      cells.push(serializeCellInline(state, cell));
-    });
-    rows.push(cells);
-  });
-
-  const columnCount = Math.max(...rows.map((row) => row.length), 1);
-  const pad = (row: string[]): string[] =>
-    row.length >= columnCount
-      ? row
-      : [...row, ...Array.from({ length: columnCount - row.length }, () => '')];
-
-  const toLine = (cells: string[]) => `| ${cells.join(' | ')} |`;
-
-  // The first row always serializes as the header because pipe tables
-  // require one; if header cells were deleted the first body row is
-  // promoted on the next parse.
-  const lines = [
-    toLine(pad(rows[0] ?? [])),
-    toLine(
-      Array.from({ length: columnCount }, (_, column) =>
-        delimiterForAlign(aligns[column] ?? null),
-      ),
-    ),
-    ...rows.slice(1).map((row) => toLine(pad(row))),
-  ];
-
-  lines.forEach((line, index) => {
-    state.write(line);
-    // The last line must not end with a newline; closeBlock owns the
-    // separation from the next block, like every other block serializer.
-    if (index < lines.length - 1) {
-      state.ensureNewLine();
-    }
-  });
-
-  state.closeBlock(node);
-}
-
-function delimiterForAlign(align: TableCellAlign | null): string {
-  switch (align) {
-    case 'left':
-      return ':---';
-    case 'center':
-      return ':---:';
-    case 'right':
-      return '---:';
-    default:
-      return '---';
-  }
-}
-
-// MarkdownSerializerState keeps its output buffer in fields that are not part
-// of the public typings. Capturing them is the only way to reuse the shared
-// inline mark/node serializers for cell content instead of a lossy
-// `textContent` dump.
-type SerializerInternals = {
-  out: string;
-  delim: string;
-  closed: PMNode | null;
-  atBlockStart: boolean;
-};
-
-function serializeCellInline(
-  state: MarkdownSerializerState,
-  cell: PMNode,
-): string {
-  const internals = state as MarkdownSerializerState & SerializerInternals;
-  const previous = {
-    out: internals.out,
-    delim: internals.delim,
-    closed: internals.closed,
-    atBlockStart: internals.atBlockStart,
-  };
-  internals.out = '';
-  internals.delim = '';
-  internals.closed = null;
-
-  let raw: string;
-  try {
-    state.renderInline(cell);
-    raw = internals.out;
-  } finally {
-    internals.out = previous.out;
-    internals.delim = previous.delim;
-    internals.closed = previous.closed;
-    internals.atBlockStart = previous.atBlockStart;
-  }
-
-  return (
-    raw
-      // Hard breaks persist as <br>, the GFM convention for line breaks in
-      // pipe-table cells; any other raw newline would terminate the row, so
-      // it flattens to a space.
-      .replace(/\\\n/g, '<br>')
-      .replace(/\n+/g, ' ')
-      // GFM expects pipes escaped everywhere in a cell, including inside
-      // inline code spans.
-      .replace(/\|/g, '\\|')
-      .trim()
-  );
 }
